+++ src/state/slices/opportunitiesSlice/resolvers/uniV2/utils.ts
@@ -2,7 +2,8 @@ import type { AssetId } from '@shapeshiftoss/caip'
 import { fromAssetId } from '@shapeshiftoss/caip'
 import type { TokenAmount } from '@uniswap/sdk'
 import type { IUniswapV2Pair } from 'contracts/__generated'
-import { ContractType, getOrCreateContractByType } from 'contracts/contractManager'
+import { getOrCreateContractByType } from 'contracts/contractManager'
+import { ContractType } from 'contracts/types'
 import { ethers } from 'ethers'
 import memoize from 'lodash/memoize'
 import type { BN } from 'lib/bignumber/bignumber'
